requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.4
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chatterbot-corpus==1.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.1
cx_Freeze==7.0.0
cx_Logging==3.2.0
debugpy==1.8.1
decorator==4.4.2
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
executing==2.0.1
filelock==3.13.4
Flask==3.0.3
Flask-RESTful==0.3.10
frozenlist==1.4.1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
investpy==1.0.8
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
Js2Py==0.74
jupyter_client==8.6.1
jupyter_core==5.7.2
lief==0.14.1
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
moviepy==1.0.3
multidict==6.0.5
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==1.25.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pefile==2023.2.7
pillow==10.3.0
pipwin==0.5.2
platformdirs==4.2.1
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
pyjsparser==2.7.1
PyJWT==2.8.0
pyparsing==3.1.2
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==3.13
pyzmq==26.0.3
regex==2024.4.28
requests==2.31.0
rsa==4.9
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.3
stack-data==0.6.3
tk==0.1.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
twilio==9.0.5
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.0
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.43.0
Wikipedia-API==0.6.0
yarl==1.9.4
